pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "AI-Driven Testing Platform"
authors = [{name = "Sentinel Team", email = "team@sentinel.dev"}]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
    "atlassian-python-api>=3.41.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "arize-phoenix>=5.0.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "pytest-httpx>=0.34.0",
    "faker>=33.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "UP", "B", "C4", "DTZ", "T20", "PT", "RET", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=sentinel --cov-report=html --cov-report=term"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel"]
